Refactor: 테스트 코드 수정
- 헬퍼 클래스 분리
- 필요없는 인코딩 테스트 케이스 제거
- HttpClient를 한번만 생성하도록 수정
