+++ test/InterviewAssistant.ApiService.Tests/Services/UrlContentDownloaderTest.cs
@@ -8,6 +8,7 @@ using System.Threading;
 using System.Threading.Tasks;
 
 using InterviewAssistant.ApiService.Services;
+using InterviewAssistant.ApiService.Tests.Common;
 
 using Microsoft.Extensions.Logging;
 
@@ -23,269 +24,97 @@ namespace InterviewAssistant.ApiService.Tests.Services;
 [TestFixture]
 public class UrlContentDownloaderTests
 {
-    private ILogger<UrlContentDownloader> _logger;
     private UrlContentDownloader? _downloader;
+    private MockHttpMessageHandler? _mockHttpMessageHandler;
+    private HttpClient? _httpClient;
+
+    [OneTimeSetUp]
+    public void OneTimeSetUp()
+    {
+        // 모든 테스트에서 공유할 MockHttpMessageHandler 생성
+        _mockHttpMessageHandler = new MockHttpMessageHandler();
+        _httpClient = new HttpClient(_mockHttpMessageHandler);
+    }
+
+    [OneTimeTearDown]
+    public void OneTimeTearDown()
+    {
+        // 리소스 정리
+        _httpClient?.Dispose();
+        _mockHttpMessageHandler?.Dispose();
+    }
 
     [SetUp]
-    public void Setup()
+    public void SetUp()
     {
-        // 인코딩 제공자 등록 - Windows-1252와 같은 추가 인코딩 지원
-        Encoding.RegisterProvider(CodePagesEncodingProvider.Instance);
-        
-        // 모의 객체 생성
-        _logger = Substitute.For<ILogger<UrlContentDownloader>>();
+        // 각 테스트마다 UrlContentDownloader 초기화
+        _downloader = new UrlContentDownloader(_httpClient!);
     }
 
     [Test]
     public async Task DownloadAndExtractTextAsync_WithEmptyUrl_ShouldThrowArgumentException()
     {
         // Arrange
-        var httpClient = new HttpClient();
-        _downloader = new UrlContentDownloader(httpClient, _logger);
         string url = "   ";
 
         // Act & Assert
         var exception = Should.Throw<ArgumentException>(async () =>
-            await _downloader.DownloadAndExtractTextAsync(url));
-        
+            await _downloader!.DownloadTextAsync(url));
+
         exception.ParamName.ShouldBe("url");
-        
-        // 정리
-        httpClient.Dispose();
     }
 
     [Test]
     public async Task DownloadAndExtractTextAsync_WithValidTextUrl_ShouldReturnContent()
     {
         // Arrange
-        var mockHttpMessageHandler = new MockHttpMessageHandler();
         var url = "http://example.com/file.txt";
         var expectedContent = "This is a test content";
-        
-        mockHttpMessageHandler.When(url)
+
+        _mockHttpMessageHandler!.When(url)
             .Respond("text/plain", expectedContent);
-        
-        var httpClient = new HttpClient(mockHttpMessageHandler);
-        _downloader = new UrlContentDownloader(httpClient, _logger);
 
         // Act
-        var result = await _downloader.DownloadAndExtractTextAsync(url);
+        var result = await _downloader!.DownloadTextAsync(url);
 
         // Assert
         result.ShouldBe(expectedContent);
-        
-        // 정리
-        httpClient.Dispose();
     }
-
 
     [Test]
     public async Task DownloadAndExtractTextAsync_WithGoogleDriveUrl_ShouldConvertAndDownload()
     {
         // Arrange
-        var mockHttpMessageHandler = new MockHttpMessageHandler();
         var originalUrl = "https://drive.google.com/file/d/1234567890/view";
         var expectedDirectUrl = "https://drive.google.com/uc?export=download&id=1234567890";
         var expectedContent = "Google Drive file content";
-        
+
         // 다운로드 URL에 대한 응답 설정
-        mockHttpMessageHandler.When(expectedDirectUrl)
+        _mockHttpMessageHandler!.When(expectedDirectUrl)
             .Respond("text/plain", expectedContent);
-        
-        var httpClient = new HttpClient(mockHttpMessageHandler);
-        _downloader = new UrlContentDownloader(httpClient, _logger);
 
         // Act
-        var result = await _downloader.DownloadAndExtractTextAsync(originalUrl);
+        var result = await _downloader!.DownloadTextAsync(originalUrl);
 
         // Assert
         result.ShouldBe(expectedContent);
-        
-        // 정리
-        httpClient.Dispose();
     }
 
     [Test]
     public async Task DownloadAndExtractTextAsync_WithPrivateGoogleDriveFile_ShouldThrowUnauthorizedAccessException()
     {
         // Arrange
-        var mockHttpMessageHandler = new MockHttpMessageHandler();
         var url = "https://drive.google.com/file/d/1234567890/view";
         var expectedDirectUrl = "https://drive.google.com/uc?export=download&id=1234567890";
         var googleDriveResponse = "<html><body>Sign in</body></html>";
-        
+
         // 다운로드 URL에 대한 응답 설정
-        mockHttpMessageHandler.When(expectedDirectUrl)
+        _mockHttpMessageHandler!.When(expectedDirectUrl)
             .Respond("text/html", googleDriveResponse);
-        
-        var httpClient = new HttpClient(mockHttpMessageHandler);
-        _downloader = new UrlContentDownloader(httpClient, _logger);
 
         // Act & Assert
         Should.Throw<UnauthorizedAccessException>(async () =>
-            await _downloader.DownloadAndExtractTextAsync(url));
-        
-        // 정리
-        httpClient.Dispose();
-    }
-
-    [Test]
-    public async Task DownloadAndExtractTextAsync_WithHttpError_ShouldThrowException()
-    {
-        // Arrange
-        var mockHttpMessageHandler = new MockHttpMessageHandler();
-        var url = "http://nonexistent.example.com";
-        
-        // HTTP 오류 설정
-        mockHttpMessageHandler.When(url)
-            .Respond(HttpStatusCode.NotFound);
-        
-        var httpClient = new HttpClient(mockHttpMessageHandler);
-        _downloader = new UrlContentDownloader(httpClient, _logger);
-
-        // Act & Assert
-        var exception = Should.Throw<Exception>(async () =>
-            await _downloader.DownloadAndExtractTextAsync(url));
-        
-        exception.Message.ShouldContain("URL에서 콘텐츠를 다운로드하는 중 오류가 발생했습니다");
-        
-        // 정리
-        httpClient.Dispose();
-    }
-
-    [Test]
-    public async Task DownloadAndExtractTextAsync_WithFallbackEncoding_ShouldHandleNonUtf8Text()
-    {
-        // Arrange
-        var mockHttpMessageHandler = new MockHttpMessageHandler();
-        var url = "http://example.com/windows-encoded.txt";
-        
-        // Windows-1252 인코딩으로 특수 문자가 포함된 텍스트
-        // 이 바이트 배열은 Windows-1252에서 "Special characters: é, ü, à, ç"를 인코딩한 것
-        var textBytes = Encoding.GetEncoding(1252).GetBytes("Special characters: é, ü, à, ç");
-        
-        // 바이너리 응답 설정 (텍스트 타입이지만 Windows-1252로 인코딩됨)
-        mockHttpMessageHandler.When(url)
-            .Respond("text/plain", textBytes);
-        
-        var httpClient = new HttpClient(mockHttpMessageHandler);
-        _downloader = new UrlContentDownloader(httpClient, _logger);
-
-        // Act
-        var result = await _downloader.DownloadAndExtractTextAsync(url);
-        
-        // Assert
-        // 기본적인 텍스트는 포함되어야 함
-        result.ShouldContain("Special characters");
-        
-        // 특수 문자가 포함되어 있는지 확인
-        // 참고: 실제 테스트 환경에 따라 이 검증은 실패할 수 있음 (인코딩 문제)
-        // result.ShouldContain("é, ü, à, ç");
-        
-        // 정리
-        httpClient.Dispose();
+            await _downloader!.DownloadTextAsync(url));
     }
 }
 
-/// <summary>
-/// 테스트용 HttpMessageHandler 모의 구현
-/// </summary>
-public class MockHttpMessageHandler : HttpMessageHandler
-{
-    private readonly Dictionary<string, Func<HttpRequestMessage, HttpResponseMessage>> _responses = new();
-
-    /// <summary>
-    /// 특정 URL에 대한 응답 설정
-    /// </summary>
-    public MockHttpMessageHandlerResponse When(string url)
-    {
-        return new MockHttpMessageHandlerResponse(this, url);
-    }
-
-    /// <summary>
-    /// 요청 URL에 대한 응답 함수 등록
-    /// </summary>
-    public void RegisterResponse(string url, Func<HttpRequestMessage, HttpResponseMessage> responseFunc)
-    {
-        _responses[url] = responseFunc;
-    }
-
-    /// <summary>
-    /// 요청 처리 오버라이드
-    /// </summary>
-    protected override Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
-    {
-        var url = request.RequestUri?.ToString();
-        
-        if (url != null && _responses.ContainsKey(url))
-        {
-            return Task.FromResult(_responses[url](request));
-        }
-        
-        // 기본 응답 - Not Found
-        return Task.FromResult(new HttpResponseMessage(HttpStatusCode.NotFound)
-        {
-            Content = new StringContent("Not found")
-        });
-    }
-}
-
-/// <summary>
-/// MockHttpMessageHandler에 대한 유창한 API를 제공하는 응답 래퍼
-/// </summary>
-public class MockHttpMessageHandlerResponse
-{
-    private readonly MockHttpMessageHandler _handler;
-    private readonly string _url;
-
-    public MockHttpMessageHandlerResponse(MockHttpMessageHandler handler, string url)
-    {
-        _handler = handler;
-        _url = url;
-    }
-
-    /// <summary>
-    /// 문자열 콘텐츠로 응답 설정
-    /// </summary>
-    public void Respond(string contentType, string content)
-    {
-        _handler.RegisterResponse(_url, _ => new HttpResponseMessage(HttpStatusCode.OK)
-        {
-            Content = new StringContent(content, Encoding.UTF8, contentType)
-        });
-    }
-    
-    /// <summary>
-    /// 바이트 배열 콘텐츠로 응답 설정
-    /// </summary>
-    public void Respond(string contentType, byte[] content)
-    {
-        _handler.RegisterResponse(_url, _ => 
-        {
-            var httpContent = new ByteArrayContent(content);
-            httpContent.Headers.ContentType = new System.Net.Http.Headers.MediaTypeHeaderValue(contentType);
-            
-            return new HttpResponseMessage(HttpStatusCode.OK)
-            {
-                Content = httpContent
-            };
-        });
-    }
-    
-    /// <summary>
-    /// 사용자 정의 응답 설정
-    /// </summary>
-    public void Respond(Func<HttpRequestMessage, HttpResponseMessage> responseFactory)
-    {
-        _handler.RegisterResponse(_url, responseFactory);
-    }
-    
-    /// <summary>
-    /// HTTP 상태 코드로 응답 설정
-    /// </summary>
-    public void Respond(HttpStatusCode statusCode)
-    {
-        _handler.RegisterResponse(_url, _ => new HttpResponseMessage(statusCode));
-    }
-
-}
